Fix crash on TCPX when setting NCCL_ALGO=TREE
The ring algorithm fallback mechanism (used when setting
NCCL_ALGO=TREE) did not take into account the fact some protocols
could be disabled.

That meant in that case, we would re-enable LL/LL128 even though we
should not.

The fix is to set the comm->ringbdw backup values to 0 when a
protocol is disabled.

Bug 4657048