Squashed RAS subsystem commit
Adds a binary executable to NCCL: bin/ncclras

Adds a new RAS debug subsystem.

Adds a new collective operations counter "collOpCount" to ncclComm, which
is not affected by the shared split communicators.

Adds the initialization of RAS to bootstrapInit and bootstrapSplit, guarded
by the NCCL_RAS_ENABLE env. variable (on by default).  The termination code
is invoked from commFree, but is currently a no-op.

Changes the initial state of comm's initState from ncclInternalError to
ncclInProgress.

Multiple changes to the generic sockets code:
- Ensure that asynchronous sockets truly don't block. Especially during
  connection establishment, the existing code could still block an async
  socket, waiting for the NCCL-level handshake to complete.
- Make it possible to avoid WARNs when the remote side closes a socket.
- Add support for one-way close (close the sending part of a socket while
  still being able to receive).
- Fix the use of SO_REUSEADDR, which had no effect.
- Make it possible to opt out of sleeping in case of ECONNREFUSED during
  connect.
- Fix the setting of TCP_NODELAY in case the socket gets recreated after a
  failed connect.