Fix potential buffer overruns related to snprintf
snprintf does not return how many bytes were written, but how many it
would have liked to have written. Some code then takes the total
bytes written (or that it wanted to write) and subtracts it from the
buffer size. This can be negative, and is interpretted as unsigned,
allowing the next iteration to overrun the buffer.