Fix NCCL_P2P_DISABLE=1 on MNNVL
MNNVL connectivity testing was being inadvertently skipped in
ncclTopoComputePaths, which calls ncclTopoCheckP2p, not p2pCanConnect.
Consequently, the test was moved from p2pCanConnect to ncclTopoCheckP2p
(which p2pCanConnect already invoked, so that path was not affected).
This required adding a comm argument to ncclTopoCheckP2p (because the
test needs to check comm->MNNVL) and corresponding changes to its
callers.

Replaced the remaining problematic instances of busId in the topo code with
rank, since busId can cause problems with MNNVL (ncclTopoCheckGdr,
ncclTopoNeedFlush).

BUG 4719744