init: fix destroy/abort comm logic
If abort is called before `initTransportsRank`, the intraComm0 pointer
is not asigned, which causes the bootstrap to not be freed.

To force the free to enter the cleanup and properly free the comm, we
setup default values for intraComm0, intraRanks, and intraRank.

This way, when the abort is called, the cleanup will happen as if the
rank is alone in its process. Once initTransportsRank has been called,
the cleanup will happen following the existing logic