transport: fix counter increment over the channels
When trying to connect the channels, we might face inconsistencies
between the data initialized  (data[p] + channelCounter) and the one
used to connect.

Assuming two channels, if channel 0 is connected, and channel 1 is not,
the connect will attempt to call transport->connect using the data
information of channel 0 ad the counter is not incremented in the case
of an already connected channel.