Squash commit for Fix Direct Read UB Bug
On the A100 platform, the P2P is read-based, which results in a bug where directRecv does not consider inpIx and directSend does not skip the send. This patch fixes this issue.

See BUG 4790817

Skip reduceCopy when UB is on and src/dst is null

See merge request nccl/nccl!558