Squash commit for Working examples of running tests on draco-rno, draco-oci and eos using run containers vs bare metal
It is now possible to build NCCL on each of these clusters with _docker/make.sh_ which will place test binaries into _build/_ directory. To illustrate how to execute the tests, we are providing working examples consisting of two parts for ease of hacking: 

* the script that wraps the test with proper env vars: _run_all_reduce_perf.sh_, which is the same for all clusters and both on-bare-metal (OBM) and in-run-container (IRC) execution options
* the _srun_ launcher in two flavors: _srun_obm.sh_ and _srun_irc.sh_ specific to each cluster

Once we narrow down our typical use cases, most of the launcher contents will be migrated to the corresponding cluster config file in _docker/_ subdirectory.

examples of running a NCCL perftest on draco-rno and eos in a run container and on bare metal

See merge request nccl/nccl!562