Fix crash when NUMA IDs are equal to -1.
This can happen when a NIC is a virtual NIC, or when linux doesn't
know which NUMA node a device is attached to.

Make sure when we compose IDs with System ID + Local ID, we ignore
the local ID bits which would spill on the system ID, like the case
of -1 which would have all bits set.

Undo previous fix at the XML level (new fix is at the topo
graph level).

Add unit test for NUMA node -1.

GitHub issue #233.
Bug 4809066
Bug 4808413