Advanced NCCL_ALGO/NCCL_PROTO per-function syntax
Allow to specify and algorithm or protocol per-function, adding a prefix
with the collective name before each token.

The current behavior of NCCL_ALGO is erratic. Users may specify it,
but in many cases we will ignore it and have a transparent fallback to
ring (when there is no algorithm, when the datatype is not supported,
etc.).
This leads to many users believing they are testing a given algorithm,
but instead using the fallback to ring.
There was also code re-enabling Ring and Tree if collnet was disabled
which interferes with new algorithms.

This change makes NCCL_ALGO follow exactly what the user requests, and
fails if there is no algo/proto available, with a better description
of what failed (hopefully so that the user realizes why it failed).

All "backup algo/proto" code is removed.

If NCCL_ALGO or NCCL_PROTO is set, a new table is printed to show the
algo/proto selection.