socket: reset fd after connect failure
When a failure is encountered in connect, the state of the socket is
unspecified. To avoid any issue and improve our robustness we reset the
fd before retrying to connect.

We also separate the O_NONBLOCK flag from the reset of the fd function
to improve in readibility and simplify cleanup in case of errors