socket: improve error handling
The errors are propagated as such up to the caller of the ncclSocketX
functions.
If the error happens in ncclSocketInit, it's unclear if the caller
should use ncclSocketClose for cleanup. Therefore we close the fd in
case of error.

if an error occurs in socketResetFd, we cleanup the newly created fd and
leave sock->fd untouched