Squash commit for ib: retry modify_qp upon error
- in case of failure in `modify_qp`, we retry if the error codes is `ETIMEDOUT`. The latter is a workaround as there is no timeout management in `ibv_modify_qp`.
- we retry the call, up to NCCL_IB_QP_RETRY_CNT times after sleeping for (NCCL_IB_QP_RETRY_SLEEP_MSEC * number of attempts) milliseconds
- the `INFO`/`WARN` message displays additional information such as the device name, the port number, the QP state, the state we transition to, and the remote and local GID
- the retry for all error codes can be enabled using `NCCL_IBV_MQP_RETRY_ALL=1`

Refactor the patch proposed in https://nvbugspro.nvidia.com/bug/4837875

ib: retry modify_qp in case of errors

The number of retries should go up to 1 min max to ensure that even long
events (20-30 secs) will be retried in case of failure.

Some error codes such as `ETIMEDOUT` should
always be retried. The others should not.
If needed, the retry for all the errors can be activated using
NCCL_IB_MQP_RETRY_ALL=1.

To help the user, we display the state of the QP, the state to which the QP had to
transition, the qp number, and the attr mask

See merge request nccl/nccl!592