Fix UDS FD leaks
For intra-node UB we added a new queryFd function to the UDS proxy
which converts a local FD to a remote FD, then returns the remote FD
to the caller (so that the caller can send that FD number to the other
side out-of-band for direct usage).

This new function works in the reverse direction of the getFd function
the UDS proxy also implements, where we ask the UDS proxy to send an FD
on its side back to the caller, converting it in the process.

Given the communication with the UDS proxy and the conversion of FDs happen
during the same ncclIpcSocketSendMsg/ncclIpcSocketRecvMsg calls, when we send
the header indicating what function we want the UDS proxy to execute, we need
to also send the local FD for conversion. In the case of a getFd request where
we don't need to convert an FD in that direction, we may be able to pass -1 as
sendFd to skip the FD conversion, but out of safety we send a dummy FD. That
FD won't be used, hence we need to free it, which was missing before.

In addition, in the queryFd case, when the UDS proxy sends the response back to
us, we don't need to convert an FD in that direction, so we should pass -1 as
the sendFd to skip the conversion (and creation).