Squash commit for ParseList accepts "not" with prefix and does better validation
ParseList accepts "not" with a prefix and does better validation. Implements better testing for ncclCommInitRank. Validates the Function/Algo/Proto table.

ParseList can now fail to parse and return ncclInvalidArgument. This can happen when: an invalid string is given for either prefix or element, or the list has an entry without a prefix which is not the first entry.

ParseList can be given NCCL_PROTO and NCCL_ALGO strings with "not" ("^") given for a prefix, not just global. For example: "NCCL_PROTO=LL;reduce:^Simple".

In order to support testing, there are new Test base classes in test/apitest/ncclCommInitRank_test.cu, two of which may have wider usage in the future and should perhaps be promoted to a common file like ncclCommon_test.cuh:

* ncclCommInitRankShelveEnvTest - A base class to allow temporary overriding of environment variables. The environment is correctly restored in TearDown, which should be called in all instances except failure during SetUp.
* ncclCommInitRankOutputTest - A base class which redirects log output to a temporary file which is read in and can be easily compared against a regular expression.

ParseList accepts "not" with prefix and does better validation

ParseList accepts "not" with a prefix and does better validation.
Implements better testing for ncclCommInitRank. Validates the
Function/Algo/Proto table.

ParseList can now fail to parse and return ncclInvalidArgument.
This can happen when: an invalid string is given for either prefix
or element, or the list has an entry without a prefix which is not
the first entry.

ParseList can be given NCCL_PROTO and NCCL_ALGO strings with "not"
("^") given for a prefix, not just global. For example:
    "NCCL_PROTO=LL;reduce:^Simple".

In order to support testing, there are new Test base classes in
test/apitest/ncclCommInitRank_test.cu, two of which may have wider
usage in the future and should perhaps be promoted to a common file
like ncclCommon_test.cuh:

* ncclCommInitRankShelveEnvTest - A base class to allow temporary
    overriding of environment variables. The environment is
    correctly restored in TearDown, which should be called in all
    instances except failure during SetUp.
* ncclCommInitRankOutputTest - A base class which redirects log
    output to a temporary file which is read in and can be easily
    compared against a regular expression.

See merge request nccl/nccl!629