Implement NCCL core profiler callback
Implement the NCCL core profiler callback used by the network plugin to
capture internal events. The `sliceSteps` is captured in the proxy
operation and used in the callback to determine the actual network step
(which no longer aligns with the algo step in NCCL). Please take a look
at commit: a9496f20c for more details.