Avoid creating a context on GPU 0 in non-blocking mode.
When adding the non-blocking feature, a couple of functions were
moved to an async thread: groupLaunch and commDestroySync.

Doing so, we forgot to remove the code which would save the current
cuda device and restore it afterwards. That code is no longer needed
in an async thread given no CUDA context is set, and restoring the
default context of GPU 0 would allocate 254MB (x8 = 2GB) on GPU 0.

For groupLaunch, in the case where it is called synchronously, we
add the save/restore code outside of the function, only in the sync
case.