Avoid unnecessary cudaGetDevice during destroy/abort
This commit makes destroy/abort wrapped with ncclGroupStart/EndInternal
to always let group op launch the commReclaim(). In this way, we can
decide to set device or not in groupEndInternal without worrying about the
cudaGetDevice issue in commReclaim().

This commit is also a follow-up of commit 4a8c8b7c