Squash commit for Harden and improve parseList tests
This adds more safety to the parseList tests. Before it was not checking that the file could be opened. It may have failed due to running on a machine where the current directory was not writeable. Now it creates the log file on /tmp, which may fix the issue. If not, there is additional checking that should prevent a crash and/or give more information on where it is failing.

Harden and improve parseList tests

See merge request nccl/nccl!654