Squash commit for Report slurm allocation and gitlab CI wait times
Report slurm allocation wait times per CSRG job within the gitlab runner scripts.

Report the overall wait and run times of gitlab runner jobs at the end of every pipeline. This is a starting point for measuring this information - we use the fields given to us by gitlab, created, started, and end times. It's not clear if we can see if a runner is overloaded or not from this info but it's at least useful to start measuring.

Report planned time

See merge request nccl/nccl!648