Add plugin version header
The plugin version header defines a schema for versioning plugins.
Currently, the plugin ID is assigned 16 bits for the network type (e.g.,
IB/SOCK/EFA/...) and 16 bits to the plugin version. This mechanism lacks
support for network types that are provided by different vendors and
event support for different plugin implementations from the same vendor.
A better schema would be to encode vendor id and sequence number along
with network type and plugin version.