Refactor net graph registration
Since legacy graph reg design would register buffer every time, for net
DMABUF registration, it would remap the buffer into BAR1 even if it is
the same buffer. It would waste the BAR1 space and cause unnecessary
registration overhead.

This commit refactors the net graph registration to work in the same
way as local registration.

New Net graph reg design (similiar to NVLS graph reg design):
1. Create static function netRegisterBuffer() for both local and graph reg.
2. For graph reg, call ncclCommRegister() to every buffer to create a local
buffer UB entry and netRegisterBuffer() to register or reuse it. cleanupNet()
is a function to deregistered the buffer when graph is destroyed.
3. For local reg, we refactor it to call netRegisterBuffer().