Refactor p2p graph reg
New design does not rely on CUDA to detect whether the buffer is
registered or not during graph reg.
Instead, the new design concept is the same as NVLS, NET, COLLNET
graph reg design.
We create a static function ipcRegisterBuffer() which is used by
both local and graph registration in order to reuse the buffer
registration as much as we can.
Similarly, for graph reg, we also call ncclCommRegister for each
buffer to increase the refCount, and deregistered through cleanupIpc()
when graph is destroyed.