Fix AR Ring Algo chunkSize assignment
chunkSize might be changed due to the remSize < loopSize, which maps
to the Simple primitive implementation. However, we should not change
the chunkSize in ringAlgo object since it can be polled by proxy
many times with different input steps where some steps can change
the chunkSize and some might not. If we change the chunkSize, it becomes
a permanent change and would cause the wrong output size and buff. To patch
it, just create a tmp var curChunkSize to represent all chunkSize.

In addition, we should use nelem to comput sliceSize. nelem should be
std::min(remSize - chunkOffset, curChunkSize).