ft_test: add cudaSetDevice before cudaMemcpy
Up until v2.23.4-1 ncclMemAlloc relied on multicast support being
present in the device to enable cuMem support. Starting with v2.24 the
logic has changed and relies on ncclIsCuMemSupported() instead. This
function detects cuMem support by checking that both

 - CU_DEVICE_ATTRIBUTE_VIRTUAL_MEMORY_MANAGEMENT_SUPPORTED and
 - CU_DEVICE_ATTRIBUTE_GPU_DIRECT_RDMA_WITH_CUDA_VMM_SUPPORTED

are set. The failure appears in v2.24 because of the change in the logic
described above. The T4 system the problem arised, does not support
multicast addresses and this causes cuMem to be disabled alltogether
in v2.23.X. With v2.24, multicast support is no longer a pre-requisite
condition for cuMem.

cuMem (introduced in cuda-10.2) allows to allocate memory and enable
access only for the peer devices that need it. This is in contrast with
cudaMalloc, where allocated memory is automatically mapped onto all the
devices that called cudaDeviceEnablePeerAccess.

With cuMem, calling cudaDeviceEnablePeerAccess is no longer needed and
only the devices for which access is set through cuMemSetAccess can
access the peer memory allocation. If cuMemSetAccess is not called, a
cudaMemcpy performed by the peer device fails with error:

cudaErrorInvalidValue (error 1) "invalid argument"

The ncclMemAlloc fails to set cuMemSetAccess for the peer device if the
devices cannot communicate using peer memory. The problem is resolved by
adding cudaSetDevice() before the cudaMemcpy call, thus letting the
originating GPU handle the copy to host memory.