Squash commit for fix ibv_retry: improve the GID log message
The current code is not complete in case of errors in `ibv_modify_qp`.

If that happens, the `qp` is unchanged, and therefore some of the information must be obtained from the user provided `attr` instead of the `ibv_query_qp` one. This is the case for `port_num` and `GID`

ibv_retry: improve the GID log message

- add the local GID index to the message
- use the user attributes instead of the QP's if the mask contains
IB_QP_AV

See merge request nccl/nccl!655