Free data array only when all connections are connected
in ncclTransportP2pSetup data[] array hold all struct ncclConnect info
of ranks which are later used to connect each other. However, we free
the data array too early so that there might be some connections that
are still under connecting. This commit move the free(data[p]) to the
end after all connections are set up.