Squash commit for Fix parseList test for GCP environment
The parseList tests are overly sensitive to the enabled algorithms
and protocols. If the test is run with NCCL_PROTO or NCCL_ALGO
already set then the code behaves correctly but the test is coded
to expect the default setting, and will fail. This change relaxes
the test to only care about the parts of the algo/proto table that
corresponds to the test being run.

Fix test for GCP environment

The parseList tests are overly sensitive to the enabled algorithms
and protocols. If the test is run with NCCL_PROTO or NCCL_ALGO
already set then the code behaves correctly but the test is coded
to expect the default setting, and will fail. This change relaxes
the test to only care about the parts of the algo/proto table that
corresponds to the test being run.

See merge request nccl/nccl!669