Use proxyState->stop as ready flag for proxy thread to exit
When CUMEM_HOST alloc is enabled, proxy thread is free to exit at any time
to free the memory. However, it still has to wait for its local gpu to complete.
So we need to first set abortFlag as 1 to indicate abort is happening, then set
proxyState->stop flag to indicate local gpu work has been done and it is safe
to release everything from proxy side.

BUG 4939036