Improve the handling of fallbacks
When searching through existing connections for a viable fallback, we
should check just the experiencingDelays flag.  Checking the socket's READY
status would prevent the connections currently being initiated from being
considered (in particular, a freshly initiated fallback for one link
wouldn't be considered for the other link, which was never the intention).

Remove all fallbacks, including the external ones, when the primary
connection becomes fully established.  External fallbacks will be re-added
if needed when handling the keep-alive messages from the peers.

Ensure the orderly shutdown of keep-alive messages over the no longer
needed external fallbacks.  When handling keep-alive messages, check for
this condition and if found, send a special nack-style keep-alive
"response" to let the peer know that it can stop sending them.  A new
"nack" flag in the keepAlive messages ensures that we don't form an
infinite back-and-forth loop of such responses.

When terminating a socket not (yet) associated with a connection, go
straight to the finalize phase as such sockets can't carry any important
messages.