Make RAS timeouts more configurable
Add NCCL_RAS_TIMEOUT_FACTOR to make it possible for a user to scale all the
RAS timeouts by an integer factor.

Support a 0 timeout value in ncclras and in the TIMEOUT command, which
disables the timeout for collective operations.  In ncclras it also
disables the fixed 1-second client-side connection timeout.

Update the docs/PLC to reflect the recent changes.