1. Fix to memory race condition in enqueue when we mix captured and non-captured launches. 2. Fix memory leaks of enqueue ncclTask*** structs since with the addition of the profiler are now allocated in the free pool.
If uncaptured work follows captured work, then its likely the uncaptured work will have
to use the hostStream to do proxyOp submission. We were erroneously freeing the proxyOps
eagerly during that submission which would lead to a race between the hostStream callback
doing the frees and the user's main thread using the free list.

The fix is to not attempt to free proxy ops eagerly, which was just an optimization to
begin with. Now proxyOps are freed in a separate pass at the conclusion of the launch ceremony
at the same time the plan struct is freed which always happens on the main thread.