Convert link->conns arrays into lists
As with the global rasSockets/rasConns/etc. structures, having them as
arrays was a mistake as it could lead to dangling pointers.

As part of the conversion refactor some of the heavy users of link->conns:

Refactor rasLinkReinitConns to leverage rasLinkAddFallback, reducing code
duplication.

Split the swiss-army-knife rasLinkUpdateConn into rasLinkConnAdd,
rasLinkConnAddExternal, and rasLinkConnUpdate.  This results in some code
duplication between the first two but makes each function vastly more
managable / easier to reason about.

Rename functions related to link->conns handling so that they all start
with rasLinkConn.