Make commHash of split communicators (more) unique
When creating split communicators, their commHash was based exclusively on
the commHash of their parent and on the passed color argument.
Consequently, repeated splits of the same parent communicator guaranteed
that different communicators would end up with the same commHash, so long
as the same color value was used (say, 0).  At least RAS needs unique
commHash values to reliably correlate communicator data from individual
ranks.

Add a new splitCount counter to ncclComm that increases on every
ncclCommSplit and use it when calculating the new commHash.  While this
extra piece of entropy certainly does not guarantee global uniqueness, it
significantly reduces the probability of collisions in trivially
reproducible circumstances.