Don't initialize intraComm0 during commAlloc
intraComm0 should be initialized after bootstrap. If we initialize it
before bootstrap, we can face the situation where some ranks pass
through the bootstrap but the other do not. In this case, if intraComm0
free its comm memory, the other ranks which pass through the bootstrap
would access the invalid intraComm0. This issue only happens when a
process manages multiple GPUs.