Report per-collective counts via RAS
This makes it possible to catch mismatched collective operation types.

Update comm->seqNumber even if no profiler plugin is loaded.

Improve the RAS output so that in case of a 0 count it reports that a rank
has "not launched any operations" rather than that it "launched up to
operation 0".

Add a new testcase with mismatched collectives.

Improve the robustness of existing triple_comm test cases with respect to
races.

Add aggregation and graph capturing to triple_comm_latecoll.