Include information on missing ranks
This improves the quality of the diagnostics for the collective
communicator info call, where now for missing ranks we can report their
complete address/pid/CUDAdev/etc.
- add nMissingRanks (and a hidden missingRanks array of struct
  rasCollCommsMissingRank) to struct rasCollComms::comm, which store the
  info on missing ranks
- add nIncompleteRanks to struct rasAuxComm which stores the count of ranks
  that failed to respond
- rename UNKNOWN status to NOCOMM to indicate that the process claims not
  be a member of a particular communicator.
- new rasGpuToString function needed by struct rasCollCommsMissingRank
- make it possible for collective-specific callbacks to modify the
  collective request before it is forwarded to peers.  This is used by
  rasCollCommsInit to provide the peers with a list of communicators for
  which missing ranks info is not needed.
- new rasCollCommsSkipMissing function needed by rasCollCommsInit

Use hostHash and pidHash in addition to commHash to disambiguate
communicators, in preparation for support for non-unique commHashes.  A new
struct rasCommId holds all three hashes.  Add the two new hashes to struct
rasPeerInfo to make it possible to look up a peer based on the missing rank
data.

Refactor a number of temporary arrays in client_support.cc that were used
for (re-)sorting.  Instead of duplicating the originals in their entirety,
use dedicated pointer+metadata structures (rasAuxPeerInfo, rasAuxCommRank)
to construct lookup tables, which are more flexible, need significantly
less memory, and should also be faster to sort.