Fixes for some MNNVL issues
The pidHash may not be unique across all the nodes on GB200-NVL
so we need to also use the hostHash for setting sameProcess.

A2A and User Buffer Registration can lead to NCCL_MAX_LOCAL_RANKS
proxy connections, so we may need to reuse a proxy connection in
ncclProxyStop() to avoid creating too many proxy connections.

BUG 5035015