Allow list in the NCCL_NET_PLUGIN
NCCL_NET_PLUGIN will accept a list instead of a single
value. This plugin load and init code will go through the
list of plugins provided and will pick the first plugin that
can be loaded and initialized successfully.
If no external plugin is specified or no external plugin can be loaded, NCCL will try to init IB and socket plugins.
When an external plugin is loaded, we disable all other external plugins.