Connect profiler proxy only if kernel events are active
Connecting the profiler proxy in init can cause a big slow down in the
ncclCommInitRankConfig function. This caused apitests on AWS to timeout
due to connection establishment times that went from ~0.7 seconds up to
over 13 seconds. Instead of unconditionally connect the profiler proxy
connect it if there is a need for it (i.e., the kernel events are
enabled).