Gracefully handle RAS cleanup during dlclose
rasTerminate() was recently added as an atexit() callback, but if
libnccl.so is loaded dynamically and then unloaded using dlclose(), that
callback is not called and the running RAS thread suddenly finds itself
attempting to execute a no longer mapped code, leading to a crash.  This
can be fixed by converting the callback to a global destructor, as the ELF
dynamic loader invokes those prior to unloading.

Also avoid the use of pthread_exit(), convenient as it may be.  glibc
apparently implements it using stack unwinding, relying on routines pulled
from libgcc_s.so.  I encountered at least one hard to diagnose case of a
hang during pthread_exit() -- not sure if it was because libnccl.so was
loaded via dlopen() or because the container had a libgcc_s.so incompatible
with libc.so -- but best to avoid such scenarios.  While fixing this,
rasLocalHandleTerminate() was renamed to rasThreadCleanup() to better
reflect its new function, and it is now invoked directly from
rasThreadMain(), replacing the old incomplete cleanup code.