Fix hang/crash on heterogeneous systems where we end up with different tree patterns.
When mixing different architectures, and in particular Ampere and
Hopper, we could end up with one node using a Tree pattern (Hopper)
and one node using Balanced Tree (Ampere).

This is not a problem for connections, but we had one line in the
tuning code adjusting the peak bandwidth depending on the graph
pattern, so different nodes could end up with a different tuning.

We cannot harmonize the graphs[]->pattern with the min/max of the
patterns from all ranks, because the pattern is used to connect rank
in ncclTopoPreset/Postset, so changing the pattern in the middle
would probably break the chains.

Instead we compute the max pattern and keep it in the comm, to
ensure all nodes will tune algorithms the same way.

Bug 5082101