hash: improve the algorithm of getHash
As discussed in
https://gitlab-master.nvidia.com/nccl/nccl/-/merge_requests/756#note_27931082
the current getHash implementation is prone to collisions.
For instance if you look at the space of all 2-byte strings, every single one
of them collides with some other 2-byte string. Yet the hash output is 8 bytes
(64-bits) so crafting an injective function ought to be easy.

The new algorithm:
* Is guaranteed injective up to 8 byte inputs.
* Spreads the entropy over the output bits so that any subset of the bits makes
  a good hash of the whole.
* Uses 2x the number of accumulator bits vs output bits. This means entropy is
  not lost when inputs contain hashes.
* Has `eatHash` for consuming multiple inputs so they don't need to be
  concatenated together, and `digestHash` for finalizing the hash value. Note
  that where inputs are broken into distinct `eatHash` calls affects the
  result. Concisely, `eatHash("ab") != eatHash("a") then eatHash("b")`

See merge request nccl/nccl!766