Reduce shared memory size for B203 GPUs
Kernel launches failed on B203 as NCCL requested more shared memory than the
device supports.
Reducing loop unroll from 8 to 6 for arch 1200 is sufficient to allow the
shared memory required for the SIMPLE protocol to fit.

Also changed the previous WAR checking for this case to be a WARN and fail.

BUG 5084471
BUG 4932193