Avoid multiple dlclose calls
NCCL plugin code shares a library handle with different sections of the code (tuner and net plugin symbols can be in the same library)
In such case we might end up calling ncclClosePluginLib twice that will eventually call dlclose on the same lib. To avoid that we can just clear all matching libhandles when calling dlclose.