Squash commit for Don't recreate unmerged devices in NIC Fusion. Perform NIC Fusion in graph_test.
Don't recreate unmerged devices in NIC Fusion:
Instead of relying on indices from the net plugin to tell us if a device is physical or virtual, we are instead relying on the keep flag. Now, all devices are marked as keep="0". Then, as they are placed (or a new one is created), the relevant nics are marked as keep="1" as NIC Fusion occurs. This removes the need to do an extra outer loop marking NICs with keep="1" or keep="0".

Perform NIC Fusion in graph_test:
To test this change, and to test NIC Fusion with much greater coverage, the functionality was added to graph_test. This should give us a lot more confidence that NIC Fusion is working as intended, isn't creating new devices, and that existing graphs are left alone.

All needed changes to the reference graphs have been made in test/unit/topo.

BUG 5101877

Got graph_test with NIC fusion functionality compiling

See merge request nccl/nccl!797