Fix a race condition during RAS initialization
This is a regression introduced in 2.26 while making RAS more resilient to
commHash conflicts.  Unfortunately, the newly added code relies on
comm->peerInfo, which is not available during bootstrap, resulting in a
crash if a RAS report is requested early on during initialization.

The commit adds a "peerInfoValid" flag to communicators.  Testing for
non-NULL peerInfo wouldn't be enough because there's a nontrivial time
window between when peerInfo is allocated and when it's initialized.

A better fix would be to change when we register new communicators with
RAS, as clearly we are doing it too early.  That would be a more disruptive
change, however, and as such it's postponed until the next NCCL release.

Issue reported on github: https://github.com/NVIDIA/nccl/issues/1669
(possibly by somebody from Meta?).

See merge request nccl/nccl!875