Fix to waitWorkFifoAvailable() in src/enqueue.cc.
We've had a bug where channels that would go unused for awhile would
make it seem the work fifo is full because their byte count was stale
and being misinterpreted due to wraparound logic. This was hanging
alltoall at moderate scale (due to them having a large fifo footprint).

Here we do away with the old approach of having channels write bytes-consumed
counters back to the host. Now we record a cuda event for the kernels completion
along with the bytes-produced count of the fifo and put those into a queue.
When we need space in the fifo we poll for the completed events to update how
much of the fifo has been consumed.

BUG 5182607

See merge request nccl/nccl!888